# apps/api-gateway/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "Con el MFA obligatorio activo, unas credenciales correctas NO devuelven una sesion: devuelven un desafio de segundo factor o un token de alta.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales invalidas"
          }
        },
        "summary": "Valida credenciales",
        "tags": [
          "Autenticacion"
        ]
      }
    },
    "/auth/mfa/verify": {
      "post": {
        "operationId": "AuthController_verifyMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMfaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Codigo MFA incorrecto"
          }
        },
        "summary": "Valida el codigo TOTP y emite la sesion",
        "tags": [
          "Autenticacion"
        ]
      }
    },
    "/auth/mfa/enroll": {
      "post": {
        "description": "Requiere el enrollmentToken de /auth/login. El factor queda inactivo.",
        "operationId": "AuthController_enrollMfa",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaEnrollmentResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Se requiere un token de alta de MFA"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Alta de MFA, paso 1: genera el secreto",
        "tags": [
          "Autenticacion"
        ]
      }
    },
    "/auth/mfa/activate": {
      "post": {
        "operationId": "AuthController_activateMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateMfaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Codigo MFA incorrecto"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Alta de MFA, paso 2: activa el factor y emite la sesion",
        "tags": [
          "Autenticacion"
        ]
      }
    },
    "/auth/sessions/current": {
      "delete": {
        "description": "El token deja de aceptarse en la siguiente peticion.",
        "operationId": "AuthController_revokeCurrentSession",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Sesion revocada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoca la sesion actual",
        "tags": [
          "Autenticacion"
        ]
      }
    },
    "/admin/ping": {
      "get": {
        "operationId": "AdminController_ping",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Acceso concedido"
          },
          "401": {
            "description": "Sin sesion valida"
          },
          "403": {
            "description": "Rol insuficiente, o token de alta de MFA"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Endpoint de referencia protegido por RBAC (rol admin)",
        "tags": [
          "Administracion"
        ]
      }
    },
    "/instruments": {
      "get": {
        "operationId": "InstrumentController_findAll",
        "parameters": [
          {
            "name": "activos",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lista los instrumentos registrados",
        "tags": [
          "Instrumentos"
        ]
      },
      "post": {
        "operationId": "InstrumentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstrumentDto"
              }
            }
          }
        },
        "responses": {
          "409": {
            "description": "El instrumento ya esta registrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registra un instrumento",
        "tags": [
          "Instrumentos"
        ]
      }
    },
    "/instruments/{symbol}": {
      "get": {
        "operationId": "InstrumentController_findOne",
        "parameters": [
          {
            "name": "symbol",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Instrumento no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtiene un instrumento por simbolo",
        "tags": [
          "Instrumentos"
        ]
      },
      "delete": {
        "description": "No lo borra: puede estar referenciado por versiones de estrategias y por operaciones ya registradas.",
        "operationId": "InstrumentController_deactivate",
        "parameters": [
          {
            "name": "symbol",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desactiva un instrumento",
        "tags": [
          "Instrumentos"
        ]
      }
    },
    "/strategies": {
      "get": {
        "operationId": "StrategyController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lista las estrategias",
        "tags": [
          "Estrategias"
        ]
      },
      "post": {
        "operationId": "StrategyController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStrategyDto"
              }
            }
          }
        },
        "responses": {
          "409": {
            "description": "Ya existe una estrategia con ese nombre"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crea una estrategia y su version inicial",
        "tags": [
          "Estrategias"
        ]
      }
    },
    "/strategies/{id}": {
      "get": {
        "operationId": "StrategyController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Estrategia no encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtiene una estrategia",
        "tags": [
          "Estrategias"
        ]
      }
    },
    "/strategies/{id}/versions": {
      "get": {
        "operationId": "StrategyController_listVersions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Historial de versiones, de la mas reciente a la mas antigua",
        "tags": [
          "Estrategias"
        ]
      },
      "post": {
        "description": "Las versiones anteriores permanecen intactas: el historial es inmutable.",
        "operationId": "StrategyController_createVersion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStrategyVersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crea una version nueva",
        "tags": [
          "Estrategias"
        ]
      }
    },
    "/strategies/{id}/status": {
      "patch": {
        "operationId": "StrategyController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStrategyStatusDto"
              }
            }
          }
        },
        "responses": {
          "409": {
            "description": "Transicion de estado no permitida"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activa, suspende o archiva una estrategia",
        "tags": [
          "Estrategias"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    }
  },
  "info": {
    "title": "roboX Capital — API Gateway",
    "description": "Punto de entrada autenticado de la plataforma. El MFA es obligatorio: unas credenciales correctas no bastan para obtener una sesion.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "analista@robox.capital"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "Nunca se registra en los logs."
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "requiresMfa": {
            "type": "boolean",
            "description": "true si hace falta un segundo factor. Con MFA obligatorio, un login correcto nunca devuelve accessToken directamente."
          },
          "requiresMfaEnrollment": {
            "type": "boolean",
            "description": "true si el usuario aun no ha configurado su segundo factor."
          },
          "challengeUserId": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador a devolver en /auth/mfa/verify."
          },
          "accessToken": {
            "type": "string",
            "description": "Solo presente si el MFA obligatorio esta desactivado por feature flag."
          },
          "enrollmentToken": {
            "type": "string",
            "description": "Token acotado, valido 10 minutos, que solo habilita /auth/mfa/enroll y /auth/mfa/activate. No da acceso a ningun recurso de negocio."
          }
        },
        "required": [
          "requiresMfa"
        ]
      },
      "VerifyMfaDto": {
        "type": "object",
        "properties": {
          "challengeUserId": {
            "type": "string",
            "format": "uuid",
            "description": "Valor devuelto por /auth/login cuando el usuario ya tiene MFA activo."
          },
          "otpCode": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6,
            "example": "123456"
          }
        },
        "required": [
          "challengeUserId",
          "otpCode"
        ]
      },
      "MfaEnrollmentResponseDto": {
        "type": "object",
        "properties": {
          "secret": {
            "type": "string",
            "description": "Secreto TOTP en base32. Mostrar una sola vez."
          },
          "otpAuthUrl": {
            "type": "string",
            "example": "otpauth://totp/roboXCapital:analista@robox.capital?secret=...&issuer=roboXCapital",
            "description": "URL para el codigo QR de la aplicacion de autenticacion."
          }
        },
        "required": [
          "secret",
          "otpAuthUrl"
        ]
      },
      "ActivateMfaDto": {
        "type": "object",
        "properties": {
          "otpCode": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6,
            "example": "123456",
            "description": "Codigo TOTP generado con el secreto entregado por /auth/mfa/enroll."
          }
        },
        "required": [
          "otpCode"
        ]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT de sesion. Enviar como Bearer en el resto de peticiones."
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "CreateInstrumentDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "example": "XAUUSD"
          },
          "name": {
            "type": "string",
            "example": "Oro contra dolar"
          },
          "assetClass": {
            "type": "string",
            "enum": [
              "metal",
              "cripto",
              "renta_variable",
              "divisa"
            ],
            "example": "metal"
          },
          "currency": {
            "type": "string",
            "example": "USD",
            "minLength": 3,
            "maxLength": 3
          },
          "exchange": {
            "type": "string",
            "example": "COMEX"
          },
          "tickSize": {
            "type": "string",
            "example": "0.01",
            "description": "Incremento minimo de precio, como cadena. No se usa un numero de coma flotante: 0.01 no es exactamente representable y el error se acumula."
          }
        },
        "required": [
          "symbol",
          "name",
          "assetClass",
          "currency",
          "tickSize"
        ]
      },
      "CreateStrategyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Momento Oro/BTC",
            "minLength": 3,
            "maxLength": 120
          },
          "description": {
            "type": "string"
          },
          "parameters": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "ventana": 20,
              "umbralEntrada": 1.5
            },
            "description": "Parametros de la version inicial. Se guardan tal cual y quedan inmutables: son la evidencia de con que configuracion se decidio cada operacion."
          },
          "instrumentIds": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "name",
          "parameters",
          "instrumentIds"
        ]
      },
      "CreateStrategyVersionDto": {
        "type": "object",
        "properties": {
          "parameters": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "ventana": 30,
              "umbralEntrada": 1.2
            }
          },
          "instrumentIds": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "parameters",
          "instrumentIds"
        ]
      },
      "ChangeStrategyStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "suspended",
              "archived"
            ],
            "description": "Transiciones permitidas: draft->active|archived, active->suspended|archived, suspended->active|archived. 'archived' es definitivo y no tiene salida."
          }
        },
        "required": [
          "status"
        ]
      }
    }
  }
}